fix: update label fallback
