+++ packages/frontend/src/components/BigNumberConfig/index.tsx
@@ -1,23 +1,12 @@
 import { Button, InputGroup } from '@blueprintjs/core';
 import { Popover2 } from '@blueprintjs/popover2';
-import { findFieldByIdInExplore, friendlyName, getFieldLabel } from 'common';
 import React, { useState } from 'react';
 import { useVisualizationContext } from '../LightdashVisualization/VisualizationProvider';
 import { InputWrapper } from './BigNumberConfig.styles';
 
 export const BigNumberConfigPanel: React.FC = () => {
-    const { bigNumberLabel, setBigNumberLabel, resultsData, explore } =
-        useVisualizationContext();
+    const { bigNumberLabel, setBigNumberLabel } = useVisualizationContext();
     const [isOpen, setIsOpen] = useState(false);
-
-    const fieldId = resultsData?.metricQuery.metrics[0];
-    const field =
-        explore && fieldId
-            ? findFieldByIdInExplore(explore, fieldId)
-            : undefined;
-    const label = field
-        ? getFieldLabel(field)
-        : fieldId && friendlyName(fieldId);
 
     return (
         <Popover2
@@ -25,7 +14,7 @@ export const BigNumberConfigPanel: React.FC = () => {
                 <InputWrapper label="Label">
                     <InputGroup
                         placeholder="Enter label"
-                        defaultValue={bigNumberLabel || label}
+                        defaultValue={bigNumberLabel}
                         onBlur={(e) => setBigNumberLabel(e.currentTarget.value)}
                     />
                 </InputWrapper>
